% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{non_zeros_voxels}
\alias{non_zeros_voxels}
\title{This is a utility function that adds two numbers}
\usage{
non_zeros_voxels(training)
}
\arguments{
\item{training}{The training data}
}
\value{
A vector of integers indicating the non-zero voxels
}
\description{
This is a utility function that adds two numbers
}
