% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_image.R
\name{normalized_image}
\alias{normalized_image}
\title{Normalize an image.}
\usage{
normalized_image(image)
}
\arguments{
\item{image}{A matrix representing an image.}
}
\value{
A matrix representing the normalized image.
}
\description{
This function normalizes an image by dividing each pixel value by the maximum possible value (255).
}
\examples{
img <- matrix(runif(100, min = 0, max = 255), ncol = 10)
normalized_img <- normalized_image(img)
}
