% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{partition_var}
\alias{partition_var}
\title{This is a utility function that subtracts two numbers}
\usage{
partition_var(training, thresholding = 0.05)
}
\arguments{
\item{training}{The training data, matrix}

\item{thresholding}{The thresholding values, prior selection probability greater than thresholding is 1, else 2}
}
\value{
The difference between the two numbers
}
\description{
This is a utility function that subtracts two numbers
}
